requirements.txt
openai
python-dotenv
pydantic
rich
pypandoc
agents